Add model support for navigating between CPU captures
This CL introduces CaptureTraceIdsIterator, which is responsible for
storing all the trace IDs of captures generated in a session, and
providing APIs to navigate across such IDs. CpuProfilerStage should
notify the iterator when new captures are created and should use its
APIs to select next/previous captures. Navigation will be triggered
using the context menu of CPU profiler and the sessions panel.

This CL changes only the model to support the navigation, and add the
unit tests to cover the feature. UI changes, including the aspect
fire/listen workflow, to follow up.

Bug: 73338399
Test: Updated existing. Added CpuCaptureTest#traceIdPassedInConstructor,
      CpuProfilerStageTest#captureNavigationChangesCaptureSelection and
      CpuProfilerStageTest#captureNavigationEnabledInSessionsWithTraces

Change-Id: I3776f1f4142e3e2a3ee31e2677f6a75cbcf15c21