Refactor rendering parsers
This CL just moves the parsers to a common package and updates the
visibility of them to restrict it as much as possible. The idea is that
refactoring of the parsers framework will be easier once they do not
have dependencies other than the minimum needed.
This CL does not contain logic changes.

Test: No logic changes. Existing tests pass
Change-Id: Iea2c456c7626188fc654ade917fa07094a52e22c