Move AppBarConfigurationDialogTest from UNRELIABLE
These test pass when running as UNRELIABLE but are flaky when running
in the DEFAULT group.
Added extra log to the rendering code to debug failures.
Used information collected during runs in the PSQ to request rendering
inside a runReadActionInSmartMode, to stop Indexing during rendering.

Change-Id: Ifae3736cc0f82aa10a53d644f62e0994d09afc73
Test-Strategy: ui.test.groups=DEFAULT
Test: AppBarConfigurationDialogTest
Bug: 72124019