Fix getOutput() to not block indefinitely. Expanded test timeout
getOutput() did a spin wait to check when the output console was
available was true. This was not ideal, since we have the Wait mechanism
for checking when a condition was true. Additionally, using a spin wait
made the wrapping Wait wrapper pointless, since the running thread would
not have a chance to return from the method to check if timeout
occurred.

This method was fixed by checking for the console only once and then
returning in a private method called "pollOutput". There is no busy
polling performed.

Additionally, the debugger test needs more time to run, so its timeout
was expanded.

Bug: 73662244
Change-Id: Idf2a34a9a6d6fb80c3fc6f7b003e5d8b9dfd0560
Fixes: 73662244
Test: all existing tests