[New PSD] Unify adding new lib dependencies to Java and Android modules.
This CL unifies how new library dependencies are added to Java and Android
modules by eliminating the need to pass the list of target build variants
from the UI code (in addition to a list of scope which specifies the scopes
for the new dependency which is passed in both cases). The change also
eliminates the possibility of getting an inconsistent state if the list of
build variants and the list of scopes does not correspond.

A new unit test is added for the simple case of adding a library dependency.

Bug: 73639646
Test: new unit tests added
Change-Id: I09e3e9a219a6678ee40daecd5f74fa985480ef6c