Fixes uses of the srcCompat attribute in the tools namespace
LayoutPsiPullParser will return srcCompat when layoutlib asks for "src"
if AppCompat is being used. This is done to prevent AppCompat from
trying to load the drawables itself since it does not support many
layoutlib features (like sample data).
When doing the replacement, we weren't looking also in the tools:
namespace so the lookup would fail if only tools:srcCompat was present.

Fixes: 73598272
Test: New regression test
Change-Id: I637d21ded8c97a4eb7a16cd2b75325a45e53b0a0