[Gradle Build Model] Flatten GradlePropertiesDslElement.
This change converts the storage of GradleDslElements in
GradlePropertiesDslElement to a single list, rather than a map of lists.
This allow us to remove the need for a key, allowing renaming to take
immediate effect. An elements name is now used for finding elements
within the list.

This will make it easier to add support for the ordering of elements
within a GradlePropertiesElement such that we can obtain greater control
of where they are placed in the file.

Bug: 72144740
Test: Unit tests updated
Change-Id: I68f1bb1216f41cc84002c8f35a0717bef26b7a3e