[Gradle Build Model] Change GradleDslMethodCall#getName
This change alters the behavior of how names are stored within
GradleDslMethodCall in the following way:

Old behavior: The name given by getName() i.e in

storeFile file('file.txt')

would be "file" and the "storeFile" would be accessible via
getStatementName(). This was in contrast with every other element where
the name of the element would by "storeFile" (i.e prop1 = "hello").

New behavior: The name given by getName() in the example above would be
"storeFile" and the method name would be obtained via
getStatementName(). This brings it more into line with other elements.
If the GradleDslMethodCall has no statement name, for example in
repositories {
  google()
}
Then getName() with return the same as getMethodName().

This change is required as a prerequisite to adding support for the
remaining SigningConfig properties as well as converting
GradleDslPropertiesModel's storage method from Maps to Lists to
instantly get the effect of renaming.

Bug: 73184484
Test: Updated existing
Change-Id: Iac27b0388eef9b556ad7e082671fb5e7967362d5