Return null from ManifestUtils.getNodeKey() if unknown key
Previously, ManifestUtils.getNodeKey() would never return
null and instead assume an element was an intent-filter
element if it didn't meet the other criteria, which would
result in an AssertionError being thrown if the element
was not an intent-filter element.

This CL changes ManifestUtils.getNodeKey() to return null
if the element doesn't meet the expected criteria.

This CL also adds some null checks to
ManifestUtils.getRecords() and refactors that method a bit.

Bug: 73329785
Test: ManifestUtilsTest.testGetNodeKeyForUnexpectedElement()
Change-Id: I972ae9dffdf91d56d7dc4a67ce3e93c5c08e13bc