NPE while creating an Instant app
After selecting Map Activity, for example, and displaying its
configuration (ConfigureTemplateParametersStep) and before pressing
"Finish", the user can go back and change the value of
"Include Android Instant App support".
This triggers two events: It resets the selected template (back to
"Empty Activity" and a notification that the "package name" has changed.
ConfigureTemplateParametersStep is listening to changes of
"package name" and triggers a re-evaluation of all Template parameters,
but uses the UI created for "Map Activity" instead of "Empty Activity".
We listen for "parameters values updated" to detect users changes on the
ConfigureTemplateParametersStep, but in this case the changes are
produced on a different "Step" and they should be ignored. When the
user re-enters ConfigureTemplateParametersStep, the UI is re-created
using the right template and displaying the right values.

Bug: 68126335
Change-Id: Ide63579197c9cada11ef7f073b734d5f739d97d9
Test-Strategy: ui.test.groups=PROJECT_WIZARD
Test: Run current UI tests. This is a edge case difficult to test.
(cherry picked from commit 7a879b6172804e6a28be70ff735931121a8aef10)