Avoid running disposed on pooled threads for unit testing
The RenderTestUtil methods used to check that the RenderTask dispose
thread is correctly terminated might fail if the thread is schedule on a
Runnable running on a pooled thread. If the pooled thread takes some
time to run, this might cause the RenderTask dispose thread to be
created AFTER the RenderTestUtil checks have ran.
This CL makes SyncLayoutlibSceneManager avoid scheduling a RenderTask
dispose in a pooled thread so it's more unit test friendly. Now
SyncLayoutlibSceneManager will put the dispose in the AWT queue since
that is correctly handled by the RenderTestUtil checks.

Test: Fixes flaky tests
Bug: 73758650
Change-Id: Id7136e1c132dedaa7563275ba21e4c7e55157843