Add a memory use test for LayoutEditor
LayoutEditorMemoryUseTest UI test looks for memory leaks for a
given scenario: open 5 layout files and navigate between them in a loop
while switching between design and text tabs.

The test uses LeakHunter to track objects that are created. If the
object count for a given class goes up with each pass, the class
is reported as a source of a leak.

The test found two leaks, each leak will be fixed in a separete commit.
ourIgnoredClasses collection stores class names on known leaks.

Test: LayoutEditorMemoryUseTest
Bug: N/A

Change-Id: Ibe0edf9113cf387e05bfd75a0fe5a1aee1760815