Add a memory use test for LayoutEditor
LayoutEditorMemoryUseTest UI test looks for memory leaks for a
given scenario: open 3 layout files and navigate between them in a loop
while switching between design and editor tabs.

The test uses LeakHunter to track objects that are created. If the
object count for a given class goes up with each pass, the class
is reported as a source of a leak.

The test found two leaks, each leak will be fixed in a separete commit.
ourIgnoredClasses collection stores class names on known leaks.

The test itself should be reliable, but is in UNRELIABLE test group due to
its heuristic nature. Ths way we can investigate any new failures without
failing the build.

Test: LayoutEditorMemoryUseTest
Bug: N/A
Change-Id: I6f7cae6dee0f7a1bcacea67f6b26ec26b62e74cf