Show pending tasks as skipped when build is cancelled.
When a build is cancelled, any remaining tasks that appear on the build
view are kept as running. This needs to change to reflect that those
tasks did not finish.

This CL calls onCancel on the request task listener when
GradleTasksExecutorImpl.ProgressIndicatorStateDelegate#onCancel is
called. Also, override
ExternalSystemTaskNotificationListenerAdapter#onCancel on GradleBuildInvoker
to have a finish event with a SkipResult, this change causes any node
that is running to appear as skipped.

Bug: 73397414
Test: Covered with existing UI test that cancells a build. Manually
verified fix.

Change-Id: Ic4d5a3310087d8e7597ebcf716269b749446da72