First pass for Energy minibar
This CL adds a minibar to the energy profiler, showing a high level
view of the events which are shown in more detail in the energy event
table view (when you do a selection).

There still needs to be further polish (e.g. design review, handling
more events, grouping similar event kinds togheter, support for doing
a selection range on the component, and maybe others), but this acts
as an initial foundation for the feature.

Fixes: 73088903
Test: Added
Change-Id: Idd5986161b77e1d20acd7e13a488829fb7e66aef