[Gradle Build Model] Add PropertyTransform tests.
This change adds tests for DefaultTransform and
SingleArgumentMethodTransform.

We also clean up the functions in SingleArgumentMethodTransform and make
the test() method more robust to enable us to use stronger assumtions in
transform().

We also PropertyUtil#createOrReplaceBasicExpression propergate the name
element, rather than using a new one.

We also fix a small bug in bind() where the parent of the created
argument element was being set to the property holder rather than the
GradleMethodCall.

Bug: N/A
Test: Unit tests added
Change-Id: Id411c342557a764469d323a3edfdb0a5a77e171a