Sample content for the What's New Assistant.
Start: Go into Tools -> Internal Actions -> Android -> Edit Studio Flags
and set the What's New assistant flag to true. Then go to
Help > What's New in Android Studio to open the assistant.

Solution 1 (Feature View): This is what you first see when you open the
panel.
Pros:
- Allows you to have a nice logo image.
- There's at least some padding between the edge of the panel and the
text. (Though, it's still very inconsistent with the first paragraph.)
- You can still include groups of tutorials as expandable nodes. This
is great if you have more complex features (i.e. the Dynamic Apps
example), or you have another set of release notes, like the Gradle
plugin.
Cons:
- The lower pane (where the expandandable lists are) can only contain
expandable lists. That is, if you simply want to have a block of text,
it still includes an expandable node (with no children).
- The point above is what caused me to include all the detailed text
and images in the header. HOWEVER, the header is not scrollable. That
is, if you have a lot of text and images, it actually pushes the nodes
out of view, and you can't scroll down to them. That's why there are
no images in this solution and the text is kept brief.

Solution 2 (Tutorial view): From the first panel you see, click through
to the tutorial in the very last node.
Pros:
- Header is scrollable! For some reason, the header in this panel behaves
differently from the one in the Feature view. So  I basically put
everything I wanted to in the header here and removed the actual
tutorial component.
Cons:
- I don't think this view allows you to have a "logo" attribute. But you
can include an "icon" image next to the title, which you can't format
differently.
- There is no padding between the edge of the panel and most of the
text.
- Obviously, you can't list groups of other tutorial views.

Some potentially easy fixes to get the best of both worlds:
- Make the header in the Feature view scrollable
(like the tutorial view)
- OR, in the Feature view, don't give feature blocks expandable nodes if
they don't include child tutorial blocks.
- Adjust the padding to align with the title/logo.

Test: none (UI text)
Fixes: 73974524
Change-Id: Icdf9f119e36efb65b967da3a93ec8c231c44a4fd