Studio side changes for CpuTiming
This change adds AtraceNodeModel to represent Atrace trace nodes wall time
as well as idle cpu time. A node is drawn for each trace event equal to the
total wall time of that event. As well as a node is drawn for each trace
event equal to the idle cpu time. As such we end up with 2 nodes for each
trace event that has idle cpu time, and 1 node for each event without idle
cpu time.

Test: Updated
Bug: 74037027
Change-Id: I0cdb3ae0985af91e472eced438051be7a28490b0