Set Profiler tab name in the initialisation routine
We were setting the tab name lazily by the time we create the UI content of
the entire window. As a result, before opening the profiler window for
the first time, we were falling back to its ID ("Android Profiler") and
that was shown when navigating to the toolbar menu View > Tool Windows.

Bug: 73814776
Test: Manual. Navigate to View > Tool Windows in the toolbar menu and
      verify the profiler entry text says "Profiler".

Change-Id: I63474fe5ee2e0860e4c88a156523b9f3cbb5285b