Refresh file system after builds with pre-3.0 AGP.
With pre-3.0 Android Gradle Plugins, AARs are exploded after
Gradle Sync, and the file watcher in IDE doesn't know when
the files becomes available, this causes "unresolved symbol"
errors.
Need to force a refresh of file system to make IDE pick up
exploded AARs.

The refresh is not needed for 3.0+ AGPs, because AARs are
exploded during Gradle Sync with 3.0+.

Fixes: 72451291
Test: GradleSyncWithOlderPluginTest.testSyncWithGradleBuildCacheUninitialized
and ProjectStructureTest.testGetAndroidPluginVersion

Change-Id: I33ed35df257fb784ec8e08909fc0c054b81188c0