[Gradle Build Model] Support GradleDslElement placement
This CL adds the required support to create elements in the correct order.
To do this we allow elements to query their parents for where they should be placed.

For the most part elements reflect the order they are stored in GradlePropertiesDslElement,
apart from when this involves non-continuous elements of the same GradleDslElementList.
See PropertyOrderTest#testAddToExistingElementList for examples of what happens here.

Bug: 72144740
Test: Added unit tests
Change-Id: I31bf177721515fef505f0b9e84032c9bf82de3d5