Dump out all registered extensions in case of an error
This is a followup to ag/Ie80249a6c0c78b2. It looks like all the plugins
are registered properly, but the root area doesn't seem to have the
extension point defined (even though the plugin is there, and its
descriptor shows the extension point). This CL just logs the full list
of defined extension points when such an issue happens: I suspect that
the root cause may be a threading issue in the platform.

Bug: 73902993
Test: existing UI tests
Change-Id: I2c2272d2c437abb3beba3bc9abd39cfc7c8de510