[AVD Mgr] Improve detection of Emulator's QEMU-2 support
Some AVD features are only available with an emulated System
Image that supports QEMU-2. (Actually, it's only one feature:
the Multi-Core CPU control.)

We detected QEMU-2 support by the presence of a 'kernel-ranchu'
file in the system-images area. Newer system images use a file
named 'kernel-ranchu-64', so this Cl will accept any file that
begins with 'kernal-ranchu'.

The UI previously included the Quick Boot control in a QEMU-2-only
group. This CL also removes that association.

Fixes: 74006401

Test: Added AvdManagerConnectionTest.testDoesSystemImageSupportQemu2

Change-Id: Ic59461c17ae7007bb71d6e16d6307c906d29cf83