Future-proof InstallerData for longer Build Number
Android Studio's build number looks currently like AI-173.xxxx, but
we are soon elongating it to something like AI-181.32.xxxx.  So the
build id part of it, 'xxxx', is not always the second number.  But
it will always be the _last_ number.  Hence replacing the older logic
with the newer logic.

Bug: 72834269, 73576836
Test: Manual QA testing on Windows
Change-Id: Ib170ef1d34932b30c5c8ec9c3a5573853b1d4cb3