[New PSD] Separate parsed and resolved dependency collections.
This CL splits the data paths for parsed and resolved dependencies in the PSD.
Two new dependency collection classes are introduced. The dependencies,
however, are still collected in the same way in both cases.

It also removes POM dependencies from the dependecies graph panel. This is
due to unavailability of required details in the parsed dependency collection.

Bug: 73999935
Test: existing
Change-Id: I947841678be7e4d8e7cdba56a80ef587e65211f4