Add GradleSyncState start calls to refresh project button
There is a new "Refresh Gradle Project" button on the Build View.
Pressing this button causes indexing to became paused forever. The cause
of that is because the refresh button calls directly to
ExternalSystemUtil#refreshProject causing all the required changes to
GradleSyncState to not be processed.

This CL adds changes to sync state when calling
AndroidGradleImportTaskNotificationListener#onStart. Also,
since sync can be re rerun from the build view button, now
GradleSyncState saves the last parameters used when calling *syncStarted
methods to be used by onStart method.

Bug: 74360539
Test: Manually tested that indexing is not kept paused.
Change-Id: I98bc70bfd813ae79c3d368112c5b5252f5f528ff