Check if project initialized when reporting unresolved dependencies
When a project is imported and some dependencies that come from the
Google Maven repository cannot be resolved (for example, an incorrect
version) we check if the repository is present on the build.gradle files
and if it is not, then we show a quick fix to add it. It is possible
that the project is not yet initialized when this check is done and in
those cases an exception is generated.

This CL prevents the exception by checking first if the project is
initialized and if it is not, adds the quickfix with a slightly modified
message to reflect that will add it if needed.

Bug: 68056337
Test: Existing tests already cover this quick fix and created
testReportNotInitialized to confirm the reporter works on unitialized
projects.

Change-Id: I44a8c7899211934810d73f299294a427c2b48f35