Clean AVD name given to DeployTargetPickerDialogFixture
If an AVD name contains a substring in parentheses in the Deploy Target
Picker Dialog, then DeployTargetPickerDialogFixture will chop off the
substring in parentheses (parentheses included) when scanning the list
of AVDs. However, the Fixture class does not perform the same String
processing on the selectDevice method's String parameter, which is the
name of the AVD to search for. If a caller passes in a name that
includes parentheses, then the fixture will never find the name in a
list of AVDs.

This is another attempt to re-introduce change
I115ac7e4a13fbf73e459252ef4ad3170832bce4d.

Test: all existing tests should pass
Change-Id: I6646457c697eba497bb7ae57e38602533e9ee88f
Fixes: 74450665