When 2 states are next to each other StateChart asserts
This change fixes the state chart assert by removing the assert,
as well as removing the mValues list in state chart. To do this
we simply reversed the map of keys to rectangles to be rectangles
to keys. If we find the creation of rectangles is slow we can use
a pool of them.

Test: Added a test to StateChartTest
Change-Id: If509d9ca87b07e9f35f6cdfbd72c50a6f46c7648
Fixes: 74499754