Fix test dependency scope
When removing dependencies from a ExcludeRoots container, first library
dependencies were being removed and after library dependencies of
dependent modules being added, while it should be the opposite. e.g.
module A depends on library C and module B, module B depends on library
C, when removing dependencies from the container, first we readd the
module dependencies and after we remove the libraries, so library C is
removed in this case.

Test: TestArtifactSearchScopesTest.testResolvedScopeForTestOnlyModuleProject
Change-Id: If4637b61aba3c00848de2ce7b30b92a8a1855247
Fixes: 74352259