Dynamic App: Fix Build and Clean project actions
When using "Build > Make Project", "Build > Clean" or "Build | Rebuild",
Android Studio should apply the action to all dynamic feature modules as
well as the base app module.

The most straightforward way to address this issue is to "fix"
ProjectStructure.getLeadModules() to return the base app module as well
as the feature modules.

Bug: 74605202
Test: Add unit test to ProjectStructureTest
Change-Id: If87161d814599728bf3ec7290f8f04ebe379cf5d