[New PSD] Simplify findElement() and fix transitive dependencies.
findElement() is never used as a memeber of PsModelCollection interface but
it signature complicates implementations. This CL removes the abstract
declaration, simplifies the implementation signtures and makes them simpler.

The CL also makes the getTransitiveDependencies() return all transitive
dependencies including different versions if they happend to be present. This
behavior will be used and tested in the following CL.

Bug: n/a
Test: existing
Change-Id: Ib04b0f69a98aa43b9ab1232acfc117be0ae74635