Ignore setting value identical to current one for Variables
In the Variables view of the PSD, do not do anything if the value of a
table cell is set a value identical to the current one. This happens in
particular when clicking in a cell and then clicking outside of it,
which Swing interprets as having edited the cell.

Bug: 74964225
Test: tests updated
Change-Id: Ibf7d9f99b0b0029065db5b8fdb2e9d17ff5a14ee