Calling sync while it is already running should not fail previous sync
We currently call ProjectSetUpTask#onFailure whenever there are sync
failures, but inside of it, we blindly change the GradleSyncState. This
causes that a new sync attempt that fails because there is a sync
already running to change the previous sync state.

This CL first confirms if the failure is because of a sync is already
running, and if that is the case, it only logs the error message but
does nothing more.

Test: New test to confirm that nothing on GradleSyncState is called when
the error is because of a sync already running.

Bug: 75005810
Change-Id: I82306ad331fa847b334a2f3d5feec94ec7afa963