Revert "Add a highlight color for multiple events in a duration."
This reverts commit d071d6347a0ef6cf129ee34e1f45d479dbf163b0.

Reason for revert: This code isn't working quite as expected.

See screenshots in action:

  Normal: https://screenshot.googleplex.com/BeLBJPtCZOQ.png
  Broken: https://screenshot.googleplex.com/1i6YCXiDTpC.png

I've also seen cases where highlight sizes are not all the same, as well as duplicate highlights:

  Broken: https://screenshot.googleplex.com/JCWx8xFfqVG.png

Usually I'd just fix this in place, but there is also ongoing work to update the statechart in a separate effort, in ag/3751964, so it's probably better if they don't have to worry about carrying around this highlight code which is probably going to change soon anyway.

We'll get a highlight solution checked back in, soon.

Change-Id: I8345692cfca9b9493e06380701f0cb394aa9b5cf