Include all events in energy details view
Before, we would only show events in the energy details
view which happened between the active timeline range.
However, for the details view, the user will expect to see
all related events, not just the time-restricted ones.

For example, if you had the following event group:

 t=100, Wake lock acquired
 t=200, Wake lock acquired
 t=300, Wake lock released
 t=400, Wake lock acquired
 t=500, Wake lock released
 t=600, Wake lock released

then no matter what the time range was, when you clicked
on a selected row in the energy table, the details view
should show all events (even if the time range was only
from, say, 250 to 350).

Change-Id: I4b92457bce0323ab3eb88dd573754c1bf044f471
Fixes: 74204071
Test: N/A, already covered by existing tests