Index-proof sync waits in BazelGuiTestProjectSystem
The current implementation of waitForProjectSyncToFinish in
BazelGuiTestProjectSystem is prone to failure if indexing happens
before or during the initial Bazel sync. This is because the method
expects Bazel sync to start within two seconds of importing the
project and assumes that indexing + the initial sync will finish
in 90 seconds by simply waiting for background tasks to finish.

This CL resolves the issue by using the output in the Bazel console
to determine when a Bazel sync is finished (rather than using
GuiTests.waitForBackgroundTasks()) and bumping the timeout for
sync start and sync finish waits up to five minutes each.

Bug: 73037396
Test: NlEditorTest, AswbGotoDeclarationTest, and BazelStartupSyncTest
      still pass.
Test-Strategy: ui.test.groups=DEFAULT

Change-Id: I4a1d88de6a8c30b2267c2ed9b838e596f0c7b7ea