Fix some update problems with a ComboBox control.
In the property inspector v2,
We were receiving model.setSelectedItem in several situations:
1) When the user selects an item in the dropdown.
2) When the ComboBox control was created.
3) When the popup was shown.
4) When focus was lost.

We need to update the property value on (1) but ignore all the
other calls. There were no good way to determine what caused the
model change.

In this CL we eliminate (2) and (4),
and explicitly perform (3) such that we know NOT to update the
property value.

In addition this CL adds logic to determine why a popup is being
closed:
1) User hits <enter>
2) User hits <escape>
3) User clicks on an item in the popup
4) User clicks on the combo control (the arrow to close the popup)
5) User clicks somewhere else

We want the value of the property to change on (1) and (3).
And the property value should be unchanged in (2), (4), and (5).

This fixes several editing problems in the ComboBox.

Bug: 73778668
Test: Existing tests pass.

Change-Id: Ia1a16dd6e79c3dba7eca6eb1ba3f91fbbd51b0d3