Handle multiple iml files for one module.
Sometimes there are multiple iml files created for one module,
for example, when opening a project created on another machine,
or when user renamed the module, or moved the module to different
location.
When this happens, DataNode in ProjectInfo has the original module
name, while IdeModifiableModelsProvider has the altered name,
for example, directory-moduleName.

In ModuleModelDataService, we need to find the Module from
IdeModifiableModelsProvider based on DataNode. Due to the reason
mentioned above, module name in DataNode and ModelsProvider can
be different, so we should not use module name as the only identifier.
Otherwide, Studio will produce error "Unable to find module with Gradle
path xxx".

ModelsProvider.findIdeModule finds the Module by all possible names,
use this API to find the Module for given DataNode.

Bug: 72798630
Test: Current tests.
Change-Id: I6705ee61f309c92427ce7226d352e1cf6213bc69