Renaming ImageDiffTest to be [Component]_[Test].png
In the future we may want to do something like Class.getName() + [Test String],
I'll leave that for us to decide later. For now this change organizes
all test by their type so they are grouped. This should make it easier
when updating test to find the files that get added/changed.

Test: ImageDiffTest
Bug: 76102933
Change-Id: I367008da397e945565c04b33368750cb2b07e49f