Fix logic around checking for redundant energy samples
In a previous version of the battery model, we used a custom
EnergySample class whose equality method only tested usage
values (e.g. network=100, cpu=500). However, when we switched
over to using protos, the equality check still compiled fine but
would always be different, because it was now testing against
timestamps as well. This caused the code to ALWAYS reject
redundant samples, meaning it would still work but would waste
memory.

Bug: N/A, local fix
Test: N/A, optimization implementation detail.
There are still tests which cover that the overall behavior of
the battery model works as expected.

Change-Id: Id7c09054db47a88fa425ef9f3a405bc147a0d969