Remove refresh button from Gradle tool window
Gradle tool window contains a "Refresh all Gradle projects" button. On
Idea, this button syncs the project by calling
ExternalSystemUtil.refreshProjects, but in Android Studio we have to
also modify GradleSyncState and do other things handled by
GradleSyncInvoker.

In order to simplify our code and remove one of the possible different
ways a user has to sync the project, it is better to remove this button
and the action associated with it (RefreshProjectAction). Also, for new
sync this action was performed by GradleProjectRefresh and now nothing
else uses this class, removing it as well.

Test: New test to confirm button is removed, and deleted tests for
RefreshProjectAction and GradleProjectRefresh.
Bug: 76099171

Change-Id: I6684e82a60a4e4e288ad2e9a8b2d8b033fcd2c79