Replace assert with Verify.verify in uitest-framework-bazel
Asserts only fail if assertions are enabled with the -ea JVM
option. It appears that we don't run UI tests with assertions enabled,
and so none of our asserts in the Bazel UI test infrastructure actually
cause the tests to fail. You can verify this by adding "assert(false)"
to the import project method and noticing that the test doesn't fail.

This CL resolves the issue by replacing all (2) asserts in
uitest-framework-bazel with calls to Verify.verify.

Fixes: 76112915
Test-Strategy: ui.test.groups=DEFAULT
Test: Verified that verify(false) causes tests to fail
      + existing Bazel UI tests still pass

Change-Id: I39818e63211f35b9f055905b8556ce8b2a61117e