Updated NullMonitorStage text for sessions
This CL switches the NullMonitorStageView to use InstructionsPanel
to display the message. This is to support adding icons in between
the text which is required for the sessions UI.

When Sessions is enabled, the message simply directs the user to the
'+' icon in the Sessions toolbar. If there are no supported device/process,
the dropdown will display the corresponding statuses.

Bug: 75961970
Test: Updated
Change-Id: Id91b936b216ac9c976c8b1c0abd37938ae165be8