Do not show dead device/process in the session's dropdown
Filter out dead devices and processes from SessionsView. There
ought to be a much bigger change to remove dead things during
the StudioProfilers update loop altogether, but that will be a future
follow-up.

This change also fixes an edge case when EndSession does not return
back the correct Session object when the device is unplugged.

Bug: 75298806
Bug: 70629566
Test: updated
Change-Id: If905f32aa2a5b26b672920b8bb31c2fce0262f83