Wait for components to be visible when importing Bazel project
The methods in ImportBazelProjectWizardFixture that are responsible
for configuring a Bazel project before importing it all assume that
the UI components that they interact with are immediately accessible.
This has caused test flakes in cases where the text fields from
the next step in the import wizard aren't visible immediately after
pushing the "next" button on the previous step.

This CL resolves the issue by adding waits to these methods to ensure
that any UI component that the fixture interacts with is available
before the fixture attempts the interaction.

Bug: 73037396
Test-Strategy: ui.test.groups=DEFAULT
Test: Bazel tests in NlEditorTest, BazelStartupSyncTest,
      and AswbGotoDeclarationTest still pass.

Change-Id: I6b9c1f79a60191e8f0784f2c665b425f48af14dc