Hide common toolbar controls when in Import Trace mode
When we're in import trace mode, we should display only zoom controls.
Navigation to other profilers through the toolbar combobox, or back to
L1 via back arrow should be disabled and so does the "Jump to live"
action.

These actions can't be trigger from anywhere, so we also disable the
"Attach to live" and "Detach from live" actions from the context menu.

Bug: 76380348
Test: Visual change. Can be manually verified by opening CPU profiler in
      import trace mode (controls shouldn't be displayed in the toolbar),
      and opening any other profiler (including CPU in normal mode). In
      that case, the toolbar should be displayed normally.

Change-Id: Ia2b3fc9884b29520385b55e87a98ce66f5e61afa