Parse imported traces into Import Trace sessions
When creating a session from an imported CPU trace file, we should parse
the file and set it as a capture in CpuProfilerStage in Import Trace
mode. If parsing fails, we warn the user through an error balloon.

The timeline boundaries should make sure the entire capture fits on the
session, so we reset the timeline when parsing is finished.

Bug: 76154264
Bug: 76190122
Test: Added CpuCaptureTest#profilerTypePassedInConstructor,
      CpuProfilerTest#referenceToTraceFilesAreSavedPerSession,
      CpuProfilerStageTest#corruptedTraceInImportTraceModeShowsABalloon,
      CpuProfilerStageTest#captureIsSetWhenOpeningStageInImportTraceMode,
      Updated other tests to reflect the changes made.

Change-Id: I20909c727579a794c1f0969a9770e338a5640b36