[Gradle Build Model] Add support for basic property reordering.
This change attempts to find the best place to put new properties
and tries its best to order the properties to match up dependencies.

The implementation performs a topological sort on the elements
contained within the Ext block whenever a new element is added,
has its value changed or is renamed.

Bug: 72144740
Test: Unit tests added.
Change-Id: I687884759a0e69f7d98aba404c2591f8266c129b