Updating FlameChart colors of Atrace Captures.
Using flame chart app color as default for all atrace events when
viewing the flame chart. Atrace is not able to distinguish between
user and system trace events. This change also unifies
getIdleCpuColor to call into getFillColor and make it darker
so we are not defing the same color in multiple places.

Test: Updating CaptureHRendererTest to validate colors.
Bug: None
Change-Id: I53f6d2491f5ad770ba504b353ba84f673473041e