Remove the "FormModel" from property inspector v2.
The "FormModel" was an attempt to keep some logical notion of a
property form which had some top level properties and a way to
move to the next line.

The only property was: "showResolvedValues" which really belong
on the highest level of the property editor. This CL removes
the notion of "resolved" values from the framework and moves it
to the Nele property model.

The logic for going to the next line is applied to the editors
after they are added to the inspector panel.

Test: Updated all unit test. Moves some tests to Nele.
Bug: 73778668
Change-Id: I44cc198be4318050b33477a918e66413885961ad