Dynamic App: Add ability to pick features to deploy
To test dynamic apps with on-demand features not installed
on the device, we add the ability for developers to pick
the list of features to deploy to the device.

Note that we keep track of the list of disabled feature (as opposed to
keeping track of the list of enabled ones) because features are enabled
by default. Since the list is stored on disk (workspace.xml) as part
of the run configuration options, it is better to keep track of
what the user has changed from the default. It also means we don't
need to update the list when a new feature is added.

Also add a HeadlessTableUI mouse listener to allow clicking on table
cells in headless mode.

Bug: 73075048
Test: Included
Change-Id: I67e67c9c4bbdb769e33431a2a5ad2b0b81da86fa