Add threads monitor to imported captures
When importing a capture, we fetch the threads from the trace file and add
them to the threads monitor so the users can view capture details per thread
by selecting one from the list.

We don't have thread states data of each thread, but we do know at what
moments within the capture there was method trace activity. Instead of
displaying a blank line corresponding to each thread, we add special states
to color the portions of the thread that contain activity, and users can use
that to select relevant portions of the capture.

Bug: 77187390
Test: ImportedTraceThreadDataSeriesTest, and
      CpuProfilerStage#threadsDataComesFromCaptureInImportTraceMode

Change-Id: I81b99e2a80b9a3a5618ec907ef84f2f8ec4ce0fb