Fix ThreeState Boolean editor.
This editor would update the property value in the model,
everytime a change value notification was received.
We dont want the editor to change anythin unless the user
changed something in the control.

In addition t turns out that the model of this editor needs
to access the "resolved" value of a property. This CL will
add this property to PropertyItem (I removed it in ag/3794343).

Bug: 73778668
Test: Unit test added.

Change-Id: Ife7db1e952d1daa8540bed8ef80d10f38338b502