Use LegacySourceProvider for testing
All tests related to namespaces need to set the model, to let the IDE
know it's supposed to work in namespaced mode. These tests don't really
care about the source sets and providers, but the Kotlin plugin does for
whatever reason. For now the TestAndroidModel just uses
LegacySourceProvider for the given facet, which seems to be enough.

Bug: N/A
Test: All touched tests no longer fail when the Kotlin plugin is enabled
      (like on the Ant buildbot).

Change-Id: I85912b7191b549bf15e2ce1a5a5275472c5af0e6