Add support for multiple inspector models.
In preparation for the motion editor:
Make the properties panel support multiple [PropertiesModel] with
corresponding [InspectorBuilder]s.

The idea is that the properties panel in Nele would show properties
from a NlComponent when a NlComponent is selected in the design
surface, and show properties for e.g. a KeyFrame when a KeyFrame is
selected in the motion editor.

The model and inspector builders should be handled completely
separately.

Bug: 73778668
Test: No tests added in this CL, may be added in a later CL.
Change-Id: If873e5dc349558f718af914d50b0c8a4bacd80ca