Apply the new dragging behavior to RelativeLayout
Dragging in RelaitveLayout will not add or remove the
constraints of component anymore.

Since RelativeLayout has supported anchors to create
constraints, it is not necessary to re-calculating the
constraints durning dragging. Instead of that, it updates
the margin values to make component can locate in the
expected position.

This behavior changing makes RelativeLayout has similar
usage of ConstraintLayout.

Some bugs of dragging component in RTL mode are also fixed.

Bug: 72422088
Test: Updated and passed RelativeLayoutHandlerKtTest
Change-Id: I171a30d8e6923755de74395c7a6ba09c746f7797