Ensure client IDE always shuts down
This change cleans up the logic for starting and stopping the client and
server: anything that does so should use JUnitServer#launchIdeAndStart
and JUnitServer#closeIdeAndStop. Helper methods that accomplish portions
of this task have been made private where possible.

Note: "stopping" the server doesn't mean exiting the process: it means
closing the connections, clearing the message pool, and stopping the
send/receive threads. It can be subsequently restarted.

When the server asks the client to shut down (via a CLOSE_IDE message),
it now waits for the client process to exit, and if it doesn't, destroys
it forcibly.

If the server does not receive any messages from the client for more
than 15 seconds (previously 180) (the client's 'KeepAliveThread' sends a
message ever 5 seconds), call closeIdeAndStop.

If the client detects that the socket connection is broken, shut down.
The server will then stop itself after hearing nothing for a bit.

Remove @Ignore from RestartIdeTest.

Bug: 77159160
Change-Id: I671c080b06316d619c5b91a14bbde4e59c01888c
Test-Strategy: ui.test.groups=TEST_FRAMEWORK,THEME
Test: RestartIdeTest