Fix missing "Gradle-aware Make" pre-launch task
Given the number of reports from Android Studio 3.1 users
losing the "Gradle-aware Make" task from the Run Configuration
during the 3.0 -> 3.1 update process, and the fact the root
cause has not be discovered yet, this CL contains code to
automatically perform the workaround we have been advocating,
which is to add back the "Gradle-aware Make" pre-launch start
the Android Run Configurations of a project.

The missing task is added when the first Gradle Sync of the session
completes (only once per project per session). When this happens,
we show a notification balloon with a link to the "Edit Configurations"
dialog so that advanced users can remove the task if the "auto-fix"
is not want they expect.

Bug: 76403139
Test: TBD
Change-Id: I3e0596d02d44f88af6ec5e8e1bd7b7289ad8ec55