Make ProjectViewFixture#changePane to wait until pane changes
ProjectViewFixture#selectPane is being used by some UI tests, this
function changes the current pane in the Project View to what the
parameters are. We have seen that sometimes this pane does not change,
causing the tests to fail.

This CL adds a wait (for up to 5 seconds) to make sure the pane was
changed before returning from changePane. Also, simplifies selectPane by
removing the viewId parameter, it is redundant as the viewId and name
should come from the same pane type.

Test: This CL fixes a testing issue.
Change-Id: Ic33daef335fd4ce3bbfa15cef171f3cfde7fe749
Bug: 77157746